pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages]
find = {where = ["src"]}

[project]
name = "hut-monitoring"
version = "1.0.0"
description = "Automatically checks hut-reservation.org for available spots and notifies via Telegram"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Astreptocoque", email = "astreptocoque@gmail.com"}
]
keywords = ["hut", "reservation", "availability", "telegram", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "playwright",
    "httpx",
    "pyyaml",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.990",
]

[project.urls]
Homepage = "https://github.com/astreptocoque/hut_monitoring"
Repository = "https://github.com/astreptocoque/hut_monitoring"
Issues = "https://github.com/astreptocoque/hut_monitoring/issues"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
